Trait pallet_farming::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
    type PalletId: Get<PalletId>;
    type WeightInfo: WeightInfo;
    type Decimal: DecimalProvider<CurrencyId>;
    type UpdateOrigin: EnsureOrigin<Self::Origin>;
    type MaxUserLockItemsCount: Get<u32>;
    type LockPoolMaxDuration: Get<Self::BlockNumber>;
    type CoolDownMaxDuration: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Currency type for deposit/withdraw assets to/from plm module

type PalletId: Get<PalletId>

Defines the pallet’s pallet id from which we can define each pool’s account id

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Decimal: DecimalProvider<CurrencyId>

Decimal provider.

type UpdateOrigin: EnsureOrigin<Self::Origin>

The origin which can create new pools and add reward.

type MaxUserLockItemsCount: Get<u32>

Specifies max amount lock item for a user

type LockPoolMaxDuration: Get<Self::BlockNumber>

Specifies upper limit of lock duration for lock pool

type CoolDownMaxDuration: Get<Self::BlockNumber>

Specifies upper limit of cool down duration for pool